requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
bioblend==0.17.0
black==22.3.0
boto==2.49.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
config==0.5.1
confusable-homoglyphs==3.2.0
distlib==0.3.4
filelock==3.4.2
idna==3.3
mypy-extensions==0.4.3
nose==1.3.7
pathspec==0.9.0
platformdirs==2.5.2
promise==2.3
protobuf==3.19.4
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
tinytag==1.8.1
tomli==2.0.1
toposort==1.7
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.13.1
whitenoise==6.0.0
